// src/app/auth/interfaces/user.interface.ts
// Generated by https://quicktype.io

import { Currency } from "src/app/dashboard/interfaces/cripto.interface";

export interface User {
  id:            number;
  firstName:     string;
  lastName:      string;
  email:         string;
  password:      string;
  country:       string;
  city:          string;
  direccion:     string;
  paymentMethod: string;
  numsTarjeta:   string;
  dateOfExpiry:  null;
  cvv:           string;
  iban:          null;
  userCurrencies:     UserCurrency[];
}

export interface UserCurrency {
  coin:     Currency;
  quantity: number;
  marketPrice: number;
}